Set up entry page: Visiting /wiki/TITLE,renders a page that displays the contents of that encyclopedia entry
If an entry is requested that does not exist, the user is presented with an error page indicating that their requested page was not found.
If the entry does exist, the user should be presented with a page that displays the content of the entry. The title of the page should include the name of the entry.